requirements.txt
streamlit
google-generativeai
requests
pdfplumber